split logic and rendering for TextBox component
