+++ src/components/layout/text-box/index.tsx
@@ -1,31 +1,9 @@
-import { FC, useRef } from "react";
-import { v4 as uidV4 } from "uuid";
-import { TaskObject } from "../../../data/task";
-import { TasksActions } from "../../../reducers/tasks/action-creators";
+import { FC } from "react";
+import { useLogic } from "./logic";
 
 export const TextBox: FC = () => {
 
-   const ref = useRef<HTMLInputElement>(null);
-
-   const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-
-      event.preventDefault();
-
-      if (!ref.current?.value) {
-         return;
-      }
-
-      const task: TaskObject = {
-         id: uidV4(),
-         name: ref.current.value,
-         time: Date.now(),
-         status: "active"
-      };
-
-      TasksActions.addTask(task);
-
-      ref.current.value = "";
-   };
+   const { ref, handleSubmit } = useLogic();
 
    return (
       <form onSubmit={handleSubmit}>
